wip(classes): new layout
