+++ .gitignore
@@ -9,4 +9,5 @@ __pycache__/
 latest_logs
 schemas.json
 build
+src/tests/tests_output/
 *.egg-info
