+++ src/opengeodeweb_viewer/rpc/mesh/schemas/display_polygon_attribute.json
@@ -1,0 +1,17 @@
+{
+  "rpc": "display_polygon_attribute",
+  "type": "object",
+  "properties": {
+    "id": {
+      "type": "string"
+    },
+    "name": {
+      "type": "string"
+    }
+  },
+  "required": [
+    "id",
+    "name"
+  ],
+  "additionalProperties": false
+}
