+++ src/opengeodeweb_viewer/rpc/viewer/schemas/apply_textures.json
