+++ src/opengeodeweb_viewer/rpc/viewer/schemas/create_object_pipeline.json
