+++ src/opengeodeweb_viewer/rpc/viewer/schemas/create_visualization.json
