+++ src/opengeodeweb_viewer/rpc/viewer/schemas/delete_object_pipeline.json
