+++ src/opengeodeweb_viewer/rpc/viewer/schemas/get_point_position.json
