+++ src/opengeodeweb_viewer/rpc/viewer/schemas/reset.json
