+++ src/opengeodeweb_viewer/rpc/viewer/schemas/reset_camera.json
