+++ src/opengeodeweb_viewer/rpc/viewer/schemas/set_opacity.json
