+++ src/opengeodeweb_viewer/rpc/viewer/schemas/set_point_size.json
@@ -1,5 +1,5 @@
 {
-  "rpc": "point_size",
+  "rpc": "set_point_size",
   "type": "object",
   "properties": {
     "id": {
@@ -9,6 +9,9 @@
       "type": "number"
     }
   },
-  "required": ["id", "size"],
+  "required": [
+    "id",
+    "size"
+  ],
   "additionalProperties": false
 }
