+++ src/opengeodeweb_viewer/rpc/viewer/schemas/set_viewer_background_color.json
