+++ src/opengeodeweb_viewer/rpc/viewer/schemas/take_screenshot.json
