+++ src/opengeodeweb_viewer/rpc/viewer/schemas/toggle_edge_visibility.json
