+++ src/opengeodeweb_viewer/rpc/viewer/schemas/toggle_object_visibility.json
