+++ src/opengeodeweb_viewer/rpc/viewer/schemas/toggle_point_visibility.json
