+++ src/opengeodeweb_viewer/rpc/viewer/schemas/update_data.json
