+++ src/tests/data/c9f110e6b28348a4a0f0cf0a924facd9.vtp
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<VTKFile type="PolyData" version="1.0" byte_order="LittleEndian" header_type="UInt32" compressor="vtkZLibDataCompressor">
+	<PolyData>
+		<Piece NumberOfPoints="1" NumberOfVerts="1">
+			<PointData>
+				<DataArray type="Float64" Name="points" format="ascii" NumberOfComponents="3" RangeMin="1" RangeMax="3">1 2 3 </DataArray>
+			</PointData>
+			<Points>
+				<DataArray type="Float64" Name="Points" format="ascii" NumberOfComponents="3" RangeMin="1" RangeMax="3">1 2 3 </DataArray>
+			</Points>
+			<Verts>
+				<DataArray type="Int64" Name="connectivity" format="ascii" RangeMin="0" RangeMax="0">0 </DataArray>
+				<DataArray type="Int64" Name="offsets" format="ascii" RangeMin="0" RangeMax="1">1 </DataArray>
+			</Verts>
+		</Piece>
+	</PolyData>
+</VTKFile>
